move all functions to root
