+++ src/functions.rs
@@ -38,3 +38,10 @@ pub fn send_input<Range: RangeBounds<usize>>(
 ) -> SendBuilder<(ProgramW, PayloadInputW<Range>, (), (), (), ())> {
     SendBuilder::input(program, payload)
 }
+
+pub fn create_program<Buffer: AsRef<[u8]>>(
+    code_id: CodeId,
+    payload: Buffer,
+) -> CreateProgramBuilder<(CodeIdW, PayloadBytesW<Buffer>, (), (), ())> {
+    CreateProgramBuilder::bytes(code_id, payload)
+}
