--- src/prog.rs
@@ -1,8 +1,0 @@
-use crate::{common::*, generated::*};
-
-pub fn create_program<Buffer: AsRef<[u8]>>(
-    code_id: CodeId,
-    payload: Buffer,
-) -> CreateProgramBuilder<(CodeIdW, PayloadBytesW<Buffer>, (), (), ())> {
-    CreateProgramBuilder::bytes(code_id, payload)
-}
